divide person.rb in diff files for each class
