+++ nameable.rb
@@ -1,0 +1,10 @@
+# nameable.rb
+
+#
+# Interface Class Nameable, for decorator design
+#
+class Nameable
+  def correct_name
+    raise NotImplementedError, "#{self.class} has not implemented method '#{__method__}'"
+  end
+end
